@@ -10,7 +10,8 @@
   "dependencies": {
     "axios": "^1.6.5",
     "core-js": "^3.8.3",
-    "vue": "^2.6.14"
+    "vue": "^3.4.15",
+    "vue-router": "^4.2.5"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
