@@ -1,11 +1,13 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
+import router from './router/router'
 import axios from 'axios';
 
 axios.defaults.baseURL = 'http://localhost:8000';
 
-Vue.config.productionTip = false
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+
+
+const app = createApp(App);
+app.use(router);
+app.mount('#app');
